@@ -4,6 +4,6 @@ class MonitorsController < ApplicationController
   newrelic_ignore
 
   def lb
-    render text: 'OK'
+    render plain: 'OK'
   end
 end
